move physics and stuff out of main.cpp does this solve the problem of bloated collision functions? no... but it may help in the future
