+++ tanks-game/tanks-game.vcxproj.filters
@@ -285,6 +285,9 @@
     <ClCompile Include="barrierpower.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="gamemainloop.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="bullet.h">
@@ -609,6 +612,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="barrierpower.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="gamemainloop.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
